refactor: reorganize shared folder, add ellipsis pipe
