+++ app/src/app/shared/ellipsis.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { EllipsisPipe } from './ellipsis.pipe';
+
+describe('EllipsisPipe', () => {
+  it('create an instance', () => {
+    const pipe = new EllipsisPipe();
+    expect(pipe).toBeTruthy();
+  });
+});
