+++ app/src/app/shared/pipes/index.ts
@@ -1,0 +1,10 @@
+import { EllipsisPipe } from './ellipsis.pipe';
+import { Provider } from '@angular/core';
+
+export {
+  EllipsisPipe
+};
+
+export const SHARED_PIPES: Provider[] = [
+  EllipsisPipe
+];
